% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{any_of_exprs}
\alias{any_of_exprs}
\title{Test that the user has entered one of several possible expressions.}
\usage{
any_of_exprs(...)
}
\arguments{
\item{...}{any number of expressions, as character strings}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if the expression the user has entered
matches any of the expressions given (as character strings) in 
the argument.
}
\examples{
\dontrun{

# Test that a user has entered either cor(x, y) or cor(y, x)
any_of_exprs('cor(x, y)', 'cor(y, x)')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}},
  \code{\link{calculates_same_value}},
  \code{\link{expr_creates_var}},
  \code{\link{expr_identical_to}}, \code{\link{expr_is_a}},
  \code{\link{expr_uses_func}},
  \code{\link{func_of_newvar_equals}},
  \code{\link{omnitest}}, \code{\link{val_has_length}},
  \code{\link{val_matches}}, \code{\link{var_is_a}}
}

