\name{bye}
\alias{bye}
\title{Exit swirl.}
\usage{
bye()
}
\description{
Swirl operates by installing a callback function which
responds to commands entered in the R console. This is how
it captures and tests answers given by the user in the R
console. Swirl will remain in operation until this callback
is removed, which is what bye() does.
}

