% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{calculates_same_value}
\alias{calculates_same_value}
\title{Test that the user's expression evaluates to a certain value.}
\usage{
calculates_same_value(expression)
}
\arguments{
\item{expression}{An expression whose value will be compared to the value
of the user's expression.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test that the value calculated by the user's expression is the same as the
value calculated by the given expression.
}
\examples{
\dontrun{
  # Test that a user's expression evaluates to a certain value
  #
  calculates_same_value('matrix(1:20, nrow=4, ncol=5)')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}},
  \code{\link{any_of_exprs}},
  \code{\link{expr_creates_var}},
  \code{\link{expr_identical_to}}, \code{\link{expr_is_a}},
  \code{\link{expr_uses_func}},
  \code{\link{func_of_newvar_equals}},
  \code{\link{omnitest}}, \code{\link{val_has_length}},
  \code{\link{val_matches}}, \code{\link{var_is_a}}
}

