% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeAssignment.R
\name{checkAssignmentFormat}
\alias{checkAssignmentFormat}
\title{checkAssignment -}
\usage{
checkAssignmentFormat(stuAns, numQuestions)
}
\arguments{
\item{orgYaml}{-}
}
\description{
checkAssignment -
}
