% Generated by roxygen2 (4.0.2): do not edit by hand
\name{course_rename}
\alias{course_rename}
\title{Rename a course folder according to its MANIFEST}
\usage{
course_rename(course_path)
}
\arguments{
\item{course_path}{Internal path to course.}
}
\description{
Rename a course folder according to its MANIFEST
}

