% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swirl.R
\name{createAssignment}
\alias{createAssignment}
\title{createAssignment}
\usage{
createAssignment(directory, assignment)
}
\arguments{
\item{directory}{- directory of yaml}

\item{assignment}{- name of assignment}
}
\description{
createAssignment
}
