% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignment.R
\name{createQuestions}
\alias{createQuestions}
\title{createQuestions}
\usage{
createQuestions(lesson, width = 80)
}
\arguments{
\item{lesson}{- data.frame of lesson}

\item{width}{- max number of characters per line (not including tags)}
}
\description{
createQuestions
}
