% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete_progress}
\alias{delete_progress}
\title{Delete a user's progress}
\usage{
delete_progress(user)
}
\arguments{
\item{user}{The user name whose progress will be deleted.}
}
\description{
Delete a user's progress
}
\examples{
\dontrun{

delete_progress("bill")
}
}

