% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{delete_swirl_option}
\alias{delete_swirl_option}
\title{Delete a global swirl option}
\usage{
delete_swirl_option(name)
}
\arguments{
\item{name}{The \code{name} of the swirl option to be deleted.}
}
\description{
Deletes the \code{name} and \code{value} of a swirl option by providing
its \code{name}.
}
\examples{
\dontrun{

delete_swirl_option("Brians_phone_number")
}
}

