% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_info.R
\name{email_admin}
\alias{email_admin}
\title{Send diagnostic email to swirl admin}
\usage{
email_admin()
}
\description{
Typing \code{email_admin()} at the prompt will attempt to open
a new email in your default browser or email client. The email
will include space for you to describe the problem you are
experiencing. It will also have the output from \code{sessionInfo},
which you should not alter.
}

