\name{expr_uses_func}
\alias{expr_uses_func}
\title{Returns TRUE if the e$expr uses the function whose
name is given as the first argument.}
\usage{
expr_uses_func(func)
}
\arguments{
  \item{x}{explain}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if the e$expr uses the function whose
name is given as the first argument.
}
\examples{
\dontrun{
}
}

