% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignment.R
\name{formatYamlForAssignment}
\alias{formatYamlForAssignment}
\title{formats Yaml file for use in assignment
createYamlForAssignment}
\usage{
formatYamlForAssignment(yaml)
}
\arguments{
\item{yaml}{- the lesson file in yaml format}
}
\description{
formats Yaml file for use in assignment
createYamlForAssignment
}
