% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{get_swirl_option}
\alias{get_swirl_option}
\title{Get a global swirl option}
\usage{
get_swirl_option(name)
}
\arguments{
\item{name}{The \code{name} of the swirl option.}
}
\description{
Returns the \code{value} of a swirl option by providing its \code{name}.
}
\examples{
\dontrun{

get_swirl_option("courses_dir")
}
}

