% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{gradeAssignment}
\alias{gradeAssignment}
\title{gradeAssignment}
\usage{
gradeAssignment(directory, assignment)
}
\arguments{
\item{directory}{- directory of yaml}

\item{assignment}{- name of assignment}
}
\description{
gradeAssignment
}
