% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{gradeFunction}
\alias{gradeFunction}
\title{gradeFunction}
\usage{
gradeFunction(current.row, e)
}
\arguments{
\item{current.row}{- directory of yaml}

\item{e}{- name of assignment
export}
}
\description{
gradeFunction
}
