% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl.R
\name{info}
\alias{info}
\title{Display a list of special commands.}
\usage{
info()
}
\description{
Display a list of the special commands, \code{bye()}, \code{play()},
\code{nxt()}, \code{skip()}, and \code{info()}.
}
\examples{
\dontrun{

| Create a new variable called `z` that contains the number 11.

> info()

| When you are at the R prompt (>):
| -- Typing skip() allows you to skip the current question.
| -- Typing play() lets you experiment with R on your own; swirl will ignore what
| you do...
| -- UNTIL you type nxt() which will regain swirl's attention.
| -- Typing bye() causes swirl to exit. Your progress will be saved.
| -- Typing info() displays these options again.

> bye()

| Leaving swirl now. Type swirl() to resume.
}
}

