\name{install_course_dropbox}
\alias{install_course_dropbox}
\title{Install a course from a zipped course directory shared on Dropbox}
\usage{
install_course_dropbox(url)
}
\arguments{
  \item{url}{URL of the shared file}
}
\description{
Install a course from a zipped course directory shared on
Dropbox
}
\examples{
\dontrun{

install_course_dropbox("https://www.dropbox.com/s/xttkmuvu7hh72vu/my_course.zip")
}
}

