\name{install_course_google_drive}
\alias{install_course_google_drive}
\title{Install a course from a zipped course directory shared on Google Drive}
\usage{
install_course_google_drive(url)
}
\arguments{
  \item{url}{URL of the shared file}
}
\description{
Install a course from a zipped course directory shared on
Google Drive
}
\examples{
\dontrun{

install_course_google_drive("https://drive.google.com/file/d/F3fveiu873hfjZZj/edit?usp=sharing")
}
}

