\name{install_course_url}
\alias{install_course_url}
\title{Install a course from a url that points to a zip file}
\usage{
install_course_url(url, type = "url", course.name = "")
}
\arguments{
  \item{url}{URL that points to a zipped course directory}
}
\description{
Install a course from a url that points to a zip file
}
\examples{
install_course_url("http://www.biostat.jhsph.edu/~rpeng/File_Hash_Course.zip")
}

