% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_from_swirl}
\alias{install_from_swirl}
\title{Install a course from the official course repository}
\usage{
install_from_swirl(course_name)
}
\arguments{
\item{course_name}{The name of the course you wish to install.}
}
\description{
We are currently maintaining a central repository of contributed
swirl courses at \url{https://github.com/swirldev/swirl_courses}.
This function provides the easiest method of installing a course
form the repository.
}
\examples{
\dontrun{

install_from_swirl("R_Programming") # Directory name

### OR ###

install_from_swirl("R Programming") # Course name
}
}

