% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_course.R
\name{InstallCourses}
\alias{InstallCourses}
\title{Installing Courses}
\description{
swirl is designed so that anyone can create interactive content
and share it with the world or with just a few people. Users can
install courses from a variety of sources using the
functions listed here. Each of these functions has its own help
file, which you can consult for more details.
}
\details{
If you're just getting started, we recommend using
\code{\link{install_from_swirl}} to install courses
from our official \href{https://github.com/swirldev/swirl_courses}{course repository}. Otherwise, check out the
help file for the relevant install function below.

You can uninstall a course from swirl at any time with
\code{\link{uninstall_course}}.

Uninstall all courses with
\code{\link{uninstall_all_courses}}.
}
\seealso{
Other InstallCourses: \code{\link{install_course_directory}};
  \code{\link{install_course_dropbox}};
  \code{\link{install_course_github}};
  \code{\link{install_course_google_drive}};
  \code{\link{install_course_url}};
  \code{\link{install_course_zip}};
  \code{\link{install_from_swirl}};
  \code{\link{uninstall_all_courses}};
  \code{\link{uninstall_course}}; \code{\link{zip_course}}
}

