% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl.R
\name{main}
\alias{main}
\title{Return to swirl's main menu.}
\usage{
main()
}
\description{
Return to swirl's main menu from a lesson in progress.
}
\examples{
\dontrun{

| The simplest way to create a sequence of numbers in R is by using
| the `:` operator. Type 1:20 to see how it works.

> main()

| Returning to the main menu...
}
}

