% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl.R
\name{nxt}
\alias{nxt}
\title{Begin the upcoming question or unit of instruction.}
\usage{
nxt()
}
\description{
This is the way to regain swirl's attention after viewing
a video or \code{play()}'ing around in the console.
}
\examples{
\dontrun{

| Create a new variable called `y` that contains the number 8.

> play()

| Entering play mode. Experiment as you please, then type nxt()
| when you ready to resume the lesson.

> 10/14
> [1] 0.7142857
> zz <- 99
> zz
> [1] 99
> nxt()

| Resuming lesson...
}
}

