\name{play}
\alias{play}
\title{Tell swirl to ignore console input for a while}
\usage{
play()
}
\description{
It is somethimes useful to play around in the R console out
of curiosity or to solidify a concept. This command will
cause swirl to remain idle, allowing the user to experiment
at will, until the command nxt() is entered.
}

