% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl.R
\name{reset}
\alias{reset}
\title{Start over on the current script question.}
\usage{
reset()
}
\description{
During a script question, this will reset the script
back to its original state, which can be helpful if you
get stuck.
}

