% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swirl.R
\name{rpt}
\alias{rpt}
\title{Repeat the previous question}
\usage{
rpt()
}
\description{
During a script question, this will move the pointer back to the previous 
row to repeat the previous question (with new values if question templates
are used.
}

