% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{select_language}
\alias{select_language}
\title{Select a language}
\usage{
select_language(append_rprofile = FALSE)
}
\arguments{
\item{append_rprofile}{If \code{TRUE} this command will append 
\code{options(swirl_language = [selected language])} to the end of your 
Rprofile. The default value is \code{FALSE}.}
}
\description{
Select a language for swirl's menus.
}

