% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{set_swirl_options}
\alias{set_swirl_options}
\title{Set global options for swirl}
\usage{
set_swirl_options(...)
}
\arguments{
\item{...}{Any options can be defined using \code{name = value}.}
}
\description{
Options can be set for swirl that have an effect on swirl's behavior. Options
are specified by the \code{name} of an option that corrsponds to the
\code{value} of that option.
}
\examples{
\dontrun{

# Install courses to the current directory
set_swirl_options(courses_dir = getwd())

# Install courses in the default course directory
set_swirl_options(courses_dir = file.path(system.file("Courses", package = "swirl")))
}
}

