% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl.R
\name{submit}
\alias{submit}
\title{Submit the active R script in response to a question.}
\usage{
submit()
}
\description{
When a swirl question requires the user to edit an R script, the
\code{submit()} function allows the user to submit their response.
}
\examples{
\dontrun{

| Create a function called f that takes one argument, x, and
| returns the value of x squared.

> submit()

| You are quite good my friend!
}
}

