\name{swirl-package}
\alias{swirl-package}
\alias{swirl}
\docType{package}
\title{
\packageTitle{swirl}
}
\description{
\packageDescription{swirl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{swirl}
\packageIndices{swirl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{swirl}

Maintainer: \packageMaintainer{swirl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
