% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeAssignment.R
\name{userFeedback}
\alias{userFeedback}
\title{userFeedback}
\usage{
userFeedback(directory, submission.file, correct)
}
\arguments{
\item{directory}{- assignment directory containing /instr/, /submission/ and /feedback/}

\item{submission.file}{- the complete path of the submission file}

\item{correct}{- a logical vector corresponding to whether the answer to a question is
TRUE or FALSE}
}
\description{
userFeedback
}
