% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{var_has_value}
\alias{var_has_value}
\title{Test that the user has created an object with a specific 
numerical value.}
\usage{
var_has_value(var, val, eps = 0)
}
\arguments{
\item{var}{an object name as a string}

\item{val}{the desired value}

\item{eps}{the error tolerance}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if the object with the given name 
has the specified value, within a certain error tolerance
}
\note{
error tolerance is needed since seemingly identical numerical 
   calculations may differ, such as 1-pnorm(3) and 
   pnorm(3, lower.tail = FALSE)
}
\examples{
\dontrun{

# Test that a user has set X = P(Z < 1.4), where Z ~ N(0,1) 
  var_has_value('X', pnorm(1.4), 1e-10)
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}},
  \code{\link{any_of_exprs}},
  \code{\link{calculates_same_value}},
  \code{\link{expr_creates_var}},
  \code{\link{expr_identical_to}}, \code{\link{expr_is_a}},
  \code{\link{expr_uses_func}},
  \code{\link{func_of_newvar_equals}},
  \code{\link{omnitest}}, \code{\link{val_has_length}},
  \code{\link{val_matches}}, \code{\link{var_is_a}}
}

