% Generated by roxygen2 (4.0.0): do not edit by hand
\name{AnswerTests}
\alias{AnswerTests}
\title{Answer Tests}
\usage{
expr_creates_var(correctName = NULL)
expr_identical_to(correct_expression)
expr_is_a(class)
expr_uses_func(func)
func_of_newvar_equals(correct_expression)
omnitest(correctExpr = NULL, correctVal = NULL, strict = FALSE)
val_has_length(len)
val_matches(regular_expression)
var_is_a(class, var_name)
}
\description{
As a swirl instructor, you can use any combination of our standard
answer tests or create your own custom answer tests to suit your
specific needs. Each of the standard answer tests listed below has
its own help file, for which you'll find a link at the bottom
of this page.
}
\section{Standard Answer Tests}{

Here is some information about the standard answer tests.
}

\section{Custom Answer Tests}{

Here is some information about writing your own custom answer tests.
}
\seealso{
Other AnswerTests: \code{\link{expr_creates_var}};
  \code{\link{expr_identical_to}}; \code{\link{expr_is_a}};
  \code{\link{expr_uses_func}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_has_length}};
  \code{\link{val_matches}}; \code{\link{var_is_a}}
}

