% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeAssignment.R
\name{checkAssignmentFormat}
\alias{checkAssignmentFormat}
\title{checkAssignmentFormat}
\usage{
checkAssignmentFormat(submission, numQuestions)
}
\arguments{
\item{submission}{- the file name of the student's submission}

\item{numQuestions}{- the actual number of questions in the assignment}
}
\value{
TRUE if the question format is correct and the assignment can be graded; returns FALSE otherwise
}
\description{
checkAssignmentFormat
}
