% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignment.R
\name{createAssignment}
\alias{createAssignment}
\title{createAssignment}
\usage{
createAssignment(yaml.file, directory, name = "assignment", width = 80,
  show = FALSE)
}
\arguments{
\item{yaml.file}{- lesson file in yaml format}

\item{directory}{- the directory to store the assignment}

\item{name}{- the name of the assignment}

\item{width}{- max number of characters per line in R_file (not including tags)}

\item{show}{- if TRUE, open the assignment (defaults to FALSE)}
}
\value{
None
}
\description{
creates an auto-gradable assignment by setting up the assignment directory 
  structure. The directory '/instr/' contains the converted yaml file and 
  the assignment as an R script; an empty '/submission/' and '/feedback/' directory are 
  also created. Method currently does not work with template-based problems.
}
