% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignment.R
\name{createAssignment}
\alias{createAssignment}
\title{createAssignment}
\usage{
createAssignment(yaml_file, R_file, width = 80, show = FALSE)
}
\arguments{
\item{yaml_file}{- lesson file in yaml format}

\item{R_file}{- name of assignment file (.R will be added)}

\item{width}{- max number of characters per line in R_file (not including tags)}

\item{show}{- if TRUE, open the assignment (defaults to FALSE)}
}
\description{
createAssignment
}
