% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{expr_creates_var}
\alias{expr_creates_var}
\title{Test that a new variable has been created.}
\usage{
expr_creates_var(correctName = NULL)
}
\arguments{
\item{correctName}{expected name of the new variable or \code{NULL}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Tests if the \code{e$expr} creates one new variable (of correct name
if given.) If so, returns \code{TRUE}.
}
\examples{
\dontrun{
# Test if the user has entered an expression which creates
# a new variable of any name.
expr_creates_var()
#
# Test if the user has entered an expression which creates
# a variable named 'myNum'
#
expr_creates_var('myNum')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}};
  \code{\link{any_of_exprs}};
  \code{\link{expr_identical_to}}; \code{\link{expr_is_a}};
  \code{\link{expr_uses_func}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_has_length}};
  \code{\link{val_matches}}; \code{\link{var_is_a}}
}

