% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{expr_identical_to}
\alias{expr_identical_to}
\title{Test that the user has entered a particular expression.}
\usage{
expr_identical_to(correct_expression)
}
\arguments{
\item{correct_expression}{the correct or expected expression as a string}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test that the user has entered an expression identical to that
given as the first argument.
}
\examples{
\dontrun{
  # Test that a user has entered a particular command
  #
  expr_identical_to('myVar <- c(3, 5, 7)')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}};
  \code{\link{any_of_exprs}};
  \code{\link{expr_creates_var}}; \code{\link{expr_is_a}};
  \code{\link{expr_uses_func}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_has_length}};
  \code{\link{val_matches}}; \code{\link{var_is_a}}
}

