% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{expr_is_a}
\alias{expr_is_a}
\title{Test that the expression itself is of a specific \code{class}.}
\usage{
expr_is_a(class)
}
\arguments{
\item{class}{expected \code{class} of the given expression}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if \code{e$expr} is of the given \code{\link{class}}.
}
\examples{
\dontrun{
# Test if the expression entered by a user is an assignment
#
expr_is_a('<-')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}};
  \code{\link{any_of_exprs}};
  \code{\link{expr_creates_var}};
  \code{\link{expr_identical_to}};
  \code{\link{expr_uses_func}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_has_length}};
  \code{\link{val_matches}}; \code{\link{var_is_a}}
}

