% Generated by roxygen2 (4.0.0): do not edit by hand
\name{expr_uses_func}
\alias{expr_uses_func}
\title{Returns TRUE if the e$expr uses the function whose
name is given as the first argument.}
\usage{
expr_uses_func(func)
}
\arguments{
\item{func}{name of the function expected to be used}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if the e$expr uses the function whose
name is given as the first argument.
}
\examples{
\dontrun{
# Test that the user has entered an expression using sd()
#
expr_uses_func('sd')
}
}

