% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{expr_uses_func}
\alias{expr_uses_func}
\title{Test that a particular function has been used.}
\usage{
expr_uses_func(func)
}
\arguments{
\item{func}{name of the function expected to be used}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if the \code{e$expr} uses the function whose
name is given as the first argument.
}
\examples{
\dontrun{
# Test that the user has entered an expression using sd()
#
expr_uses_func('sd')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}};
  \code{\link{any_of_exprs}};
  \code{\link{expr_creates_var}};
  \code{\link{expr_identical_to}}; \code{\link{expr_is_a}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_has_length}};
  \code{\link{val_matches}}; \code{\link{var_is_a}}
}

