% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{func_of_newvar_equals}
\alias{func_of_newvar_equals}
\title{Test the result of a computation applied to a specific (user-named) 
variable created in a previous question.}
\usage{
func_of_newvar_equals(correct_expression)
}
\arguments{
\item{correct_expression}{expression expected to be applied}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Tests the result of a computation such as \code{mean(newVar)} applied
to a specific variable created in a previous question and
saved behind the scenes as \code{e$newVar}.
}
\examples{
\dontrun{
# Test if user has taken the mean of a variable created
# in an earlier question.
#
func_of_newvar_equals('mean(newVar)')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}},
  \code{\link{any_of_exprs}},
  \code{\link{calculates_same_value}},
  \code{\link{expr_creates_var}},
  \code{\link{expr_identical_to}}, \code{\link{expr_is_a}},
  \code{\link{expr_uses_func}}, \code{\link{omnitest}},
  \code{\link{val_has_length}}, \code{\link{val_matches}},
  \code{\link{var_is_a}}
}

