% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swirl.R
\name{info}
\alias{info}
\title{Display a list of special commands.}
\usage{
info()
}
\description{
Display a list of the special commands, \code{bye()}, \code{play()}, 
\code{nxt()}, \code{skip()}, and \code{info()}.
}

