% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course_directory}
\alias{install_course_directory}
\title{Install a course from a course directory}
\usage{
install_course_directory(path)
}
\arguments{
\item{path}{The path to the course directory.}
}
\description{
Install a course from a course directory
}
\examples{
\dontrun{

install_course_directory("~/Desktop/my_course")
}
}
\seealso{
Other InstallCourses: \code{\link{InstallCourses}};
  \code{\link{install_course_dropbox}};
  \code{\link{install_course_github}};
  \code{\link{install_course_google_drive}};
  \code{\link{install_course_url}};
  \code{\link{install_course_zip}};
  \code{\link{install_from_swirl}};
  \code{\link{uninstall_all_courses}};
  \code{\link{uninstall_course}}; \code{\link{zip_course}}
}

