% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course_dropbox}
\alias{install_course_dropbox}
\title{Install a course from a zipped course directory shared on Dropbox}
\usage{
install_course_dropbox(url, multi = FALSE)
}
\arguments{
\item{url}{URL of the shared file}

\item{multi}{The user should set to \code{TRUE} if the zipped directory contains multiple courses. The default value is \code{FALSE}.}
}
\description{
Install a course from a zipped course directory shared on Dropbox
}
\examples{
\dontrun{

install_course_dropbox("https://www.dropbox.com/s/xttkmuvu7hh72vu/my_course.zip")
}
}
\seealso{
Other InstallCourses: \code{\link{InstallCourses}},
  \code{\link{install_course_directory}},
  \code{\link{install_course_github}},
  \code{\link{install_course_google_drive}},
  \code{\link{install_course_url}},
  \code{\link{install_course_zip}},
  \code{\link{install_from_swirl}},
  \code{\link{uninstall_all_courses}},
  \code{\link{uninstall_course}}, \code{\link{zip_course}}
}

