% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_course_google_drive}
\alias{install_course_google_drive}
\title{Install a course from a zipped course directory shared on Google Drive}
\usage{
install_course_google_drive(url, multi = FALSE)
}
\arguments{
\item{url}{URL of the shared file}

\item{multi}{The user should set to \code{TRUE} if the zipped directory contains multiple courses. The default value is \code{FALSE}.}
}
\description{
Install a course from a zipped course directory shared on Google Drive
}
\examples{
\dontrun{

install_course_google_drive("https://drive.google.com/file/d/F3fveiu873hfjZZj/edit?usp=sharing")
}
}

