% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course_google_drive}
\alias{install_course_google_drive}
\title{Install a course from a zipped course directory shared on Google Drive}
\usage{
install_course_google_drive(url, multi = FALSE)
}
\arguments{
\item{url}{URL of the shared file}

\item{multi}{The user should set to \code{TRUE} if the zipped directory contains multiple courses. The default value is \code{FALSE}.}
}
\description{
Install a course from a zipped course directory shared on Google Drive
}
\examples{
\dontrun{

install_course_google_drive("https://drive.google.com/file/d/F3fveiu873hfjZZj/edit?usp=sharing")
}
}
\seealso{
Other InstallCourses: \code{\link{InstallCourses}};
  \code{\link{install_course_directory}};
  \code{\link{install_course_dropbox}};
  \code{\link{install_course_github}};
  \code{\link{install_course_url}};
  \code{\link{install_course_zip}};
  \code{\link{install_from_swirl}};
  \code{\link{uninstall_all_courses}};
  \code{\link{uninstall_course}}; \code{\link{zip_course}}
}

