\name{install_course_zip}
\alias{install_course_zip}
\title{Install a course from a zipped course folder}
\usage{
install_course_zip(path)
}
\arguments{
  \item{path}{The path to the zipped course.}
}
\description{
Install a course from a zipped course folder
}
\examples{
install_course_zip("~/Desktop/my_course.zip")
}

