\name{install_from_swirl}
\alias{install_from_swirl}
\title{Install a course from swirl's official list of courses}
\usage{
install_from_swirl(course_name)
}
\arguments{
  \item{course_name}{The name of the course you wish to
  install.}
}
\description{
Install a course from swirl's official list of courses
}
\examples{
\dontrun{

install_from_swirl("Set Theory")
}
}

