% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_course.R
\name{uninstall_course}
\alias{uninstall_course}
\title{Uninstall a course}
\usage{
uninstall_course(course_name)
}
\arguments{
\item{course_name}{Name of course to be uninstalled}
}
\description{
Uninstall a course
}
\examples{
\dontrun{

uninstall_course("Linear Regression")
}
}
\seealso{
Other InstallCourses: \code{\link{InstallCourses}};
  \code{\link{install_course_directory}};
  \code{\link{install_course_dropbox}};
  \code{\link{install_course_github}};
  \code{\link{install_course_google_drive}};
  \code{\link{install_course_url}};
  \code{\link{install_course_zip}};
  \code{\link{install_from_swirl}};
  \code{\link{uninstall_all_courses}};
  \code{\link{zip_course}}
}

