% Generated by roxygen2 (4.0.0): do not edit by hand
\name{val_has_length}
\alias{val_has_length}
\title{Test the the length of e$val is that given by the first argument}
\usage{
val_has_length(len)
}
\arguments{
\item{len}{expected length of the variable created by a user}
}
\value{
TRUE or FALSE
}
\description{
Test the the length of e$val is that given by the first argument
}
\examples{
\dontrun{
# Test that the user has created a varible of length 10
#
val_has_length(10)
}
}

