% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{val_has_length}
\alias{val_has_length}
\title{Test that the value of the expression has a particular \code{length}.}
\usage{
val_has_length(len)
}
\arguments{
\item{len}{expected length of the variable created by a user}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test the the \code{\link{length}} of \code{e$val} is that given by the
first argument.
}
\examples{
\dontrun{
# Test that the user has created a varible of length 10
#
val_has_length(10)
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}};
  \code{\link{any_of_exprs}};
  \code{\link{expr_creates_var}};
  \code{\link{expr_identical_to}}; \code{\link{expr_is_a}};
  \code{\link{expr_uses_func}};
  \code{\link{func_of_newvar_equals}};
  \code{\link{omnitest}}; \code{\link{val_matches}};
  \code{\link{var_is_a}}
}

