% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerTests2.R
\name{val_matches}
\alias{val_matches}
\title{Test that the user's expression matches a regular expression.}
\usage{
val_matches(regular_expression)
}
\arguments{
\item{regular_expression}{a regular expression which user value should match}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if \code{as.character(e$val)} matches the regular
expression given as the first argument.
}
\examples{
\dontrun{
  # Test that a user has entered a value matching
  # '[Cc]ollege [Ss]tudents' or has selected it 
  # in a multiple choice question.
  #
  val_matches('[Cc]ollege [Ss]tudents')
}
}
\seealso{
Other AnswerTests: \code{\link{AnswerTests}},
  \code{\link{any_of_exprs}},
  \code{\link{calculates_same_value}},
  \code{\link{expr_creates_var}},
  \code{\link{expr_identical_to}}, \code{\link{expr_is_a}},
  \code{\link{expr_uses_func}},
  \code{\link{func_of_newvar_equals}},
  \code{\link{omnitest}}, \code{\link{val_has_length}},
  \code{\link{var_is_a}}
}

